Fixed error in markers, was int8 for 2D watershedding

+++ cloudtrack/watershedding.py
@@ -63,7 +63,7 @@ def watershedding_3D(track,field_in,threshold=3e-3,target='maximum',level=None,c
             raise ValueError('unknown type of target')
 
             raise ValueError('unknown type of target')
-        markers = np.zeros_like(unmasked).astype(np.int32)
+        markers = np.zeros_like(unmasked)
         for index, row in tracks_i.iterrows():
              markers[:,int(row['hdim_1']), int(row['hdim_2'])]=row.particle
         markers[~unmasked]=0
@@ -142,13 +142,13 @@ def watershedding_2D(track,field_in,threshold=0,target='maximum',compactness=0,m
             data_i_watershed=data_i
         else:
             raise ValueError('unknown type of target')
-        markers = np.zeros_like(unmasked).astype(np.int32)
+        markers = np.zeros_like(unmasked)
         for index, row in tracks_i.iterrows():
             markers[int(row['hdim_1']), int(row['hdim_2'])]=row.particle
         markers[~unmasked]=0
 
         if method=='watershed':
-            res1 = watershed(data_i_watershed,markers.astype(np.int8), mask=unmasked,compactness=compactness)
+            res1 = watershed(data_i_watershed,markers.astype(np.int32), mask=unmasked,compactness=compactness)
 #        elif method=='random_walker':
 #            #res1 = random_walker(Mask, markers,mode='cg')
 #              res1=random_walker(data_i_watershed, markers.astype(np.int8), beta=130, mode='bf', tol=0.001, copy=True, multichannel=False, return_full_prob=False, spacing=None)
